pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bact-mdr-profiler"
version = "1.0.0"
description = "Causal co-resistance discovery with hypergraph multi-drug resistance modelling"
requires-python = ">=3.10"
license = {text = "MIT"}
readme = "README.md"
authors = [
    {name = "Maciej Kochanowski", email = "m.kochanowski@piwet.pulawy.pl"}
]
keywords = ["antimicrobial resistance", "MDR", "XDR", "causal discovery", "PC algorithm", "hypergraph", "Streptococcus suis", "bioinformatics"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Intended Audience :: Science/Research",
]
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "statsmodels>=0.14",
    "joblib>=1.3",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "networkx>=3.1",
]

[project.urls]
Homepage = "https://github.com/MK-vet/bact-mdr-profiler"
Repository = "https://github.com/MK-vet/bact-mdr-profiler"
"Bug Tracker" = "https://github.com/MK-vet/bact-mdr-profiler/issues"

[project.optional-dependencies]
gui = ["marimo>=0.9", "plotly>=5.18", "matplotlib>=3.7"]
dev = ["pytest>=7", "pytest-cov>=4", "ruff>=0.4"]
all = ["bact-mdr-profiler[gui,dev]"]

[project.scripts]
bact-mdr-profiler = "bactmdrprofiler.cli:main"
bact-mdr-profiler-dashboard = "bactmdrprofiler.dashboard_launcher:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
